Moved map Parser class to new map package
