+++ bubolo/test/bubolo/integration/CollisionTestApplication.java
@@ -13,9 +13,9 @@ import bubolo.AbstractGameApplication;
 import bubolo.GameApplication;
 import bubolo.audio.Audio;
 import bubolo.graphics.Graphics;
+import bubolo.map.Parser;
 import bubolo.net.Network;
 import bubolo.net.NetworkSystem;
-import bubolo.util.Parser;
 import bubolo.world.World;
 import bubolo.world.entity.concrete.Tank;
