+++ bubolo/test/bubolo/integration/TankRespawnTestApplication.java
@@ -15,9 +15,9 @@ import bubolo.audio.Audio;
 import bubolo.graphics.Graphics;
 import bubolo.graphics.SpawnSpriteViewer;
 import bubolo.graphics.Sprites;
+import bubolo.map.Parser;
 import bubolo.net.Network;
 import bubolo.net.NetworkSystem;
-import bubolo.util.Parser;
 import bubolo.world.World;
 import bubolo.world.entity.Entity;
 import bubolo.world.entity.concrete.Spawn;
