+++ bubolo/test/bubolo/integration/TankTerrainSpeedTest.java
@@ -9,9 +9,9 @@ import org.json.simple.parser.ParseException;
 import bubolo.GameApplication;
 import bubolo.audio.Audio;
 import bubolo.graphics.Graphics;
+import bubolo.map.Parser;
 import bubolo.net.Network;
 import bubolo.net.NetworkSystem;
-import bubolo.util.Parser;
 import bubolo.world.World;
 import bubolo.world.entity.concrete.Tank;
